refactor(desktop): split the single 214-LOC lib.rs into inject + key modules
funput-desktop was one 214-LOC lib.rs (over the repo's 150-LOC-per-file bar).
Split it by its two pure concerns, with tests moving alongside their code.
Pure internal reorganization — no public API or behavior change.

- inject.rs: InjectPlan + plan_inject (what a hook+inject shell emits)
- key.rs: Mods / KeyEvent / KeyKind / classify (what a keystroke means)
- lib.rs: crate doc + module tree + re-exports

The crate root re-exports the same items (KeySource, InjectPlan, plan_inject,
KeyEvent, KeyKind, Mods, classify), so the Windows shell (funput-windows) is
untouched — verified via `cargo check --target x86_64-pc-windows-gnu`.
